+++ modules/elementary_functions/src/include/CorrelationCoefficients.hpp
@@ -30,15 +30,8 @@
 //=============================================================================
 namespace Nelson {
 //=============================================================================
-/**
- * Element-wise power.  A .^ B
- */
-//=============================================================================
 NLSELEMENTARY_FUNCTIONS_IMPEXP ArrayOf
-DotPower(ArrayOf& A, ArrayOf& B, bool& needToOverload);
-//=============================================================================
-NLSELEMENTARY_FUNCTIONS_IMPEXP ArrayOf
-DoPowerTwoArgFunction(ArrayOf A, ArrayOf B);
+CorrelationCoefficients(const ArrayOf& A, bool& needToOverload);
 //=============================================================================
 } // namespace Nelson
 //=============================================================================