+++ modules/elementary_functions/src/include/Variance.hpp
@@ -30,15 +30,9 @@
 //=============================================================================
 namespace Nelson {
 //=============================================================================
-/**
- * Element-wise power.  A .^ B
- */
+NLSELEMENTARY_FUNCTIONS_IMPEXP
+ArrayOf
+Variance(const ArrayOf& A, int w, int dim, bool& needToOverload);
 //=============================================================================
-NLSELEMENTARY_FUNCTIONS_IMPEXP ArrayOf
-DotPower(ArrayOf& A, ArrayOf& B, bool& needToOverload);
-//=============================================================================
-NLSELEMENTARY_FUNCTIONS_IMPEXP ArrayOf
-DoPowerTwoArgFunction(ArrayOf A, ArrayOf B);
-//=============================================================================
-} // namespace Nelson
+}
 //=============================================================================